Backend style errors fixed
