+++ src/test/java/com/socialmedia/controllers/UserControllerTest.java
@@ -1,4 +1,4 @@
-package com.socialmedia.Controllers;
+package com.socialmedia.controllers;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
